Move 'LocalFeedLoader' to its own file in production
