+++ EssentialFeed/EssentialFeed/Feed Cache/LocalFeedLoader.swift
@@ -1,0 +1,46 @@
+//
+//  LocalFeedLoader.swift
+//  EssentialFeed
+//
+//  Created by Nayem, Mufakkharul | Nil | GSSD on 2025/05/27.
+//
+
+import Foundation
+
+public class LocalFeedLoader {
+    private let store: FeedStore
+    private let currentDate: () -> Date
+    
+    public init(store: FeedStore, currentDate: @escaping () -> Date) {
+        self.store = store
+        self.currentDate = currentDate
+    }
+    
+    public func save(_ items: [FeedItem], completion: @escaping (Error?) -> Void) {
+        store.deleteCachedFeed { [weak self] error in
+            guard let self else { return }
+            
+            if let error {
+                completion(error)
+            } else {
+                cache(items, with: completion)
+            }
+        }
+    }
+    
+    private func cache(_ items: [FeedItem], with completion: @escaping (Error?) -> Void) {
+        store.insert(items, timestamp: currentDate()) { [weak self] error in
+            guard self != nil else { return }
+            
+            completion(error)
+        }
+    }
+}
+
+public protocol FeedStore {
+    typealias DeletionCompletion = (Error?) -> Void
+    typealias InsertionCompletion = (Error?) -> Void
+    
+    func deleteCachedFeed(completion: @escaping DeletionCompletion)
+    func insert(_ items: [FeedItem], timestamp: Date, completion: @escaping InsertionCompletion)
+}
